+++ ci/docs/build.sh
@@ -34,7 +34,7 @@ logger "Activate conda env..."
 source activate gdf
 conda install -c nvidia -c rapidsai -c rapidsai-nightly -c numba -c conda-forge \
     cudatoolkit=${CUDA_REL} cudf=$CUDF_VERSION rmm=$RMM_VERSION nvgraph networkx \
-    python-louvain sphinx sphinx_rtd_theme numpydoc sphinxcontrib-websupport \
+    python-louvain scikit-learn sphinx sphinx_rtd_theme numpydoc sphinxcontrib-websupport \
     nbsphinx ipython pandoc=\<2.0.0 recommonmark
 
 pip install sphinx-markdown-tables