+++ cpp/tests/sssp/sssp_test.cu
@@ -193,7 +193,7 @@ class Tests_SSSP : public ::testing::TestWithParam<SSSP_Usecase> {
     ASSERT_TRUE((typeid(DistType) == typeid(float)) || (typeid(DistType) == typeid(double)));
     if (param.type_ == RMAT) {
       // This is size_t due to grmat_gen which should be fixed there
-      // TODO rmat is disabled
+      // FIXME: rmat is disabled
       return;
     } else if (param.type_ == MTX) {
       MaxVType m, k;
@@ -395,7 +395,7 @@ TEST_P(Tests_SSSP, CheckFP64_NO_RANDOM_DIST_PREDS)
   run_current_test<int, int, double, false, true, true>(GetParam());
 }
 
-// TODO: There might be some tests that are done twice (MTX that are not patterns)
+// FIXME: There might be some tests that are done twice (MTX that are not patterns)
 TEST_P(Tests_SSSP, CheckFP32_RANDOM_DIST_NO_PREDS)
 {
   run_current_test<int, int, float, true, true, false>(GetParam());